Added Lookup objects, Field.get_lookup()
The Lookup object API is still totally unfininshed. There is just one
core Lookup object, BackwardsCompatLookup(lookup_name).

The sql/where.py, field and lookup interaction is a total mess
currently (whereas it was just a mess before).

Even with these limitations it is now possible to define custom
lookups for fields. It should be possible to define fields & lookups
like:
  - 'hstorefield__somekey__lte=val'
  - 'xmlfield__xpath=my_complex_xpath'
  - 'arrayfield__idx0=someval'
and so on.